Updated backend with enhanced CRUD operations and streamlined user management. Commit details:
Implemented basic CRUD operations in controllers for Medicine, MedicineSymptom, Order, OrderItem, Symptom, and User. Pending further testing beyond AuthController.
Introduced logging in BackEndApplication.java to signal application startup, aiding in debugging and operational verification.
Consolidated Admin and User classes into a single User entity to simplify user management, resulting in the removal of the standalone Admin JPA class.
Adjusted application.yml logging level from DEBUG to INFO post-resolution of CORS policy and entity column data type mismatch issues, optimizing log output.
Resolved entity column data type incompatibility by standardizing on Byte across related entities, eliminating errors and enhancing database schema consistency.
Removed separate authentication entity in favor of using the User class for authentication processes, accompanied by an authentication service class for handling logic, simplifying the authentication flow.

+++ back-end/src/main/java/com/capstone/backend/jpa/MedicineSymptom.java
@@ -58,31 +58,31 @@ class MedicineSymptomId implements Serializable { // CHANGED FROM PUBLIC TO DEFA
 	private static final long serialVersionUID = 1L;
 
     @Column(name = "medicine_id")
-    private Integer medicineId;
+    private Byte medicineId;
 
     @Column(name = "symptom_id")
-    private Integer symptomId;
+    private Byte symptomId;
 
     public MedicineSymptomId() {}
 
-    public MedicineSymptomId(Integer medicineId, Integer symptomId) {
+    public MedicineSymptomId(Byte medicineId, Byte symptomId) {
         this.medicineId = medicineId;
         this.symptomId = symptomId;
     }
 
-    public Integer getMedicineId() {
+    public Byte getMedicineId() {
         return medicineId;
     }
 
-    public void setMedicineId(Integer medicineId) {
+    public void setMedicineId(Byte medicineId) {
         this.medicineId = medicineId;
     }
 
-    public Integer getSymptomId() {
+    public Byte getSymptomId() {
         return symptomId;
     }
 
-    public void setSymptomId(Integer symptomId) {
+    public void setSymptomId(Byte symptomId) {
         this.symptomId = symptomId;
     }
